updated ai logic (using groq rest api now)
